Move utils functions to utils
