+++ src/auxiliary_functions.c
@@ -1,5 +1,6 @@
 #include "../include/auxiliary.h"
 #include "../include/globals.h"
+#include "utils.h"
 
 extern symbol* head_symbol;
 extern symbol* head_entries;
@@ -10,22 +11,6 @@ extern code_word code_table[TABLE_SIZE];
 extern int D;
 extern char* file_name;
 
-
-/*This function gets a string that represent a line content and skip the first white spaces in it. It return a pointer to the first non white space character. */
-char* delete_first_spaces(char* line)
-{
-	while (*line == ' ' || *line == '\t')
-		line++;
-	return line;
-}
-
-/*This function gets an extension of a file name and add/change extentsion in the global file_name variable.*/
-void add_extension_2_file_name(char* extension)
-{
-	int i;
-	for (i = 0; file_name[i] != '.'; i++);
-	strcpy((file_name + i), extension);
-}
 
 void free_data_table()
 {
@@ -111,26 +96,6 @@ void free_externals_list()
 		free(temp);
 	}
 	head_externals = NULL;
-}
-
-/*This function gets a string that represent a line content and initalizes it.*/
-void clean_line(char* line)
-{
-	int i;
-	for (i = 0; i < MAX_LINE_LENGTH; i++)
-	{
-		line[i] = '\0';
-	}
-}
-
-/*This function gets a string that represent a label content and initalizes it.*/
-void clean_label_name(char* label)
-{
-	int i;
-	for (i = 0; i < MAX_LABEL_LENGTH; i++)
-	{
-		label[i] = '\0';
-	}
 }
