+++ src/first_pass.c
@@ -5,6 +5,7 @@
 #include "../include/constants_tables.h"
 #include "../include/globals.h"
 #include "../include/status_codes.h"
+#include "utils.h"
 
 /*Operation code.Reliable only when the action is valid*/
 int opcode;
