+++ src/second_pass.c
@@ -4,6 +4,7 @@
 #include "../include/second_pass.h"
 #include "../include/logger.h"
 #include "output.h"
+#include "utils.h"
 
 
 /**
